feat: video hashing

+++ commands/reddit/imghash.ts
@@ -4,6 +4,13 @@ import imghash from 'imghash'
 import leven from 'leven'
 import fs from 'fs'
 import request from 'request'
+import ffmpeg from '@ffmpeg-installer/ffmpeg'
+import ffprobe from '@ffprobe-installer/ffprobe'
+import videoHash from 'video-hash'
+const vHash = new videoHash({
+    ffmpegPath: ffmpeg.path,
+    ffprobePath: ffprobe.path
+})
 const upvote = 'https://toppng.com/uploads/preview/reddit-clipart-icon-reddit-upvote-transparent-11562895696nryk8bvsps.png'
 const downvote = 'https://www.vhv.rs/dpng/d/127-1278380_reddit-downvote-transparent-hd-png-download.png'
 export default {
@@ -16,18 +23,29 @@ export default {
     async isRepost(): Promise<Boolean> {
         return true 
     },
-    async callback(message:Message, Image:string, name: string) : Promise<void> {
+    async callback(message:Message, Image:string, name: string, type: string) : Promise<void> {
         const hashs:string[] = db.get('shitpost.hash') || []
         request.get(Image)
             .on('error', console.error)
             .pipe(fs.createWriteStream('tmp/'+name))
             .on('close', async () => {
-                const hash1 = await imghash.hash("tmp/"+name);
+                var video:any 
+                if(type == 'video/mp4')
+                video = vHash.video("tmp/"+name)
+                const hash1 = video ? await video.hash({strength: 1}) : await imghash.hash("tmp/"+name)
+                if(!hash1 || typeof hash1 !== 'string') return
                 fs.unlink("tmp/"+name, (err) => {})
+                db.set('shitpost.hash', hashs.filter((hash) => {
+                    console.log()
+                    return hash && typeof hash === "string"
+                }))
                 for(const hash2 of hashs) {
+                    if(!hash2 || hash2 == 'undefined') {
+                        continue
+                    }
                     const distance = leven(hash1, hash2);
                     console.log(`Distance between images is: ${distance}`);
-                    if (distance <= 5) {
+                    if (distance <= 8) {
                         return message.reply('Repost atma lan ammmmmcik')
                     }
                 }
